+++ src/booking/business_logic/booking_service.py
@@ -12,16 +12,14 @@ class BookingService:
         self.booking_repo = BookingRepository()
         self.MOVIE_SERVICE_URL = os.getenv("MOVIE_SERVICE_URL", "http://127.0.0.1:5001")
         self.RABBITMQ_HOST = os.getenv("RABBITMQ_HOST", "localhost")
-        self.RABBITMQ_URL = os.getenv("RABBITMQ_URL", f"amqp://guest:guest@{self.RABBITMQ_HOST}:5672/")
+        self.RABBITMQ_URL = os.getenv("RABBITMQ_URL")
 
     def send_ticket_email(self, booking_id, email, seat_number, movie_title):
         try:
-            if self.RABBITMQ_HOST != 'localhost':
-                 params = pika.ConnectionParameters(host=self.RABBITMQ_HOST, port=5672)
-            elif 'localhost' in self.RABBITMQ_URL:
-                 params = pika.ConnectionParameters('localhost')
-            else:
+            if self.RABBITMQ_URL:
                  params = pika.URLParameters(self.RABBITMQ_URL)
+            else:
+                 params = pika.ConnectionParameters(host=self.RABBITMQ_HOST, port=5672)
             
             connection = pika.BlockingConnection(params)
             channel = connection.channel()
@@ -108,8 +106,11 @@ class BookingService:
     def cancel_all_bookings_for_showtime(self, showtime_id):
         return self.booking_repo.delete_all_bookings_by_showtime(showtime_id)
 
-    def create_seats(self, showtime_id, rows_count, seats_per_row):
-        self.booking_repo.create_seats_realistic(showtime_id, rows_count, seats_per_row)
+    def create_seats(self, showtime_id, rows_count, seats_per_row=None):
+        if seats_per_row is None:
+            self.booking_repo.create_seats(showtime_id, rows_count)
+        else:
+            self.booking_repo.create_seats_realistic(showtime_id, rows_count, seats_per_row)
 
     def get_seats(self, showtime_id):
         # We need to get room info from Movie Service to initialize if needed